feat(devops): Always run backend tests (#8111)
# Motivation

We want to use the status check `backend-tests-pass` to allow (or not)
the merging of a PR. However, we have the issue that if the PR does not
modify the files specified in the event trigger (in this case all
backend-related files), the workflow does not run, and the status check
is not elaborated.

To bypass this issue, we specify the files to watch for each job of the
workflow, so that the workflow always runs, but not necessarily "work"
if the files are not changed.

# Changes

- Use custom action `check-backend-changes` in each job of the workflow
that runs the backend tests.

# Tests

The CI workflow ran in this same PR correctly.