fix(devops): Specify the output for action that checks backend changes (#8121)
# Motivation

The `any` keyword for action `dorny/paths-filter` is apparently reserved
to check if ANY of the files changed, independently of the filters.

So, to avoid any confusions, we specify another keyword for the output
of the action and for the filter.

# Tests

Tested in PR https://github.com/dfinity/oisy-wallet/pull/8120 where the
files changed were more that the backend-related ones:

<img width="2068" height="496" alt="Screenshot 2025-08-03 at 06 53 18"
src="https://github.com/user-attachments/assets/56ebbc21-f11e-4ea0-aef4-5d825e137d53"
/>