fix(devops): Correct import of env variable if there is no secret (#8117)
# Motivation

When an env variable is included in the CI workflow that deploys the
frontend, it is being fetched from GitHub secrets. However, if it is not
set there, defaults to an empty string, raising automatically an error
during building when parsing with JSON.

This is still acceptable, since it would be inconsistent to tell the
deployment workflow to use a variable that we have not set. If we don't
want to set it, we simply can remove it from the workflow and it will
default to `false` during building time.

However, it is useful to know why the building fails during the CI. So,
in this PR we make the env variable parser more verbose.

# Changes

- Throw an error in `parseBoolEnvVar` when the input is an empty string.
Explain the reason why it would fail and what to do to solve it.

# Tests

I added a mock env variable in this PR, adding it to all the CI
workflows and the frontend, but without setting it among the GItHub
secrets. As expected, the deployment failed (see
https://github.com/dfinity/oisy-wallet/actions/runs/16690657397/job/47247923104),
as usual. But this time it provided more info:

<img width="2282" height="1162" alt="Screenshot 2025-08-02 at 08 16 47"
src="https://github.com/user-attachments/assets/a87bab4e-3c83-4605-8df4-f6f869355b2c"
/>