feat(frontend): extended address contacts derived (#8138)
# Motivation

Since the original contact address schema does not include any unique
identifiers, we need to create an extended store where we manually add
`id` to each address. This is needed for the AI assistant feature - we
will be hiding the address values before feeding the data to LLM, and
then parsing the response back by using these generated ids.