chore(frontend): fix main.js loader paths (#8172)
# Motivation

> TL;DR: we got lucky

I implemented a 404 page in the Juno Console. One thing led to another,
and I ended up having to hash the filenames of `main.js` loaders
(because 404.html and index.html both live in the same folder, at the
root). That’s when I noticed an issue on mainnet: the loaders were no
longer being loaded, and the frontend was crashing!

The root cause was that the loader was always being fetched from the
root directory. This wouldn’t normally be an issue—*if* all loaders had
the same content and were named `main.js` but, since I started hashing
the filenames… yolo.

Long story short, I fixed the issue by using a relative path for the
loaders (in this [PR](https://github.com/junobuild/juno/pull/1826)).
While doing that, I realized the same issue exists in OISY. Fortunately,
because all the loaders are currently identical, it hasn’t caused any
problems yet. Still, for future-proofing, it’s probably best to address
it now.

# Notes

I double-checked NNS dapp. It’s fine there, thanks to a PR from
@bitdivine back in February 2023.

# Changes

- In `build.csk.mjs`, when `injectScriptLoader`, calculate a parent
relative path and use it to load the `main.js` script loader from the
related HTML page.