feat(frontend): improves display logic for welcome modal (#8194)
# Motivation

We only want to display the `welcome` modal if no other modal is opened.
Currently we have the issue, that if the activation of a `code` (Vip,
Gold or Clickbee) takes some time, we display the welcome modal and
after a short time, it closes and displays the win modal. This should be
avoided.

# Changes

- improves modal handling

# Tests

**before:**


https://github.com/user-attachments/assets/dab3b615-161a-4c80-accc-53c1695cb90e



**after:**


https://github.com/user-attachments/assets/6ab79365-a502-4f87-83c7-382686d24d93

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>