feat(frontend): added withdrawUserUnusedBalance for ICP Swap (#8247)
# Motivation

We encountered an issue with ICP Swap where, after a swap operation, a
portion of both the sourceToken and destinationToken may remain in the
pool under the user's balance. This PR introduces functionality to
detect and withdraw any such unused token balances left in the pool.

# Changes

- Added withdrawUserUnusedBalance utility for checking and withdrawing
unused user pool balances.
- Updated SwapParams type definitions to support new logic.
- Integrated withdrawUserUnusedBalance to handle manual withdrawal
requests and after a successful swap.

# Tests

Covered new logic with tests