feat(frontend): NFTs and collections filter and sorting utils (#8282)
# Motivation

We also need some util functions to sort and filter the NFTs and NFT
collections.

We use a collator and dynamic comparator function, which is not needed
since we only have one comparator right now. But its planned to add
additional sorting types so we prepare for that.

# Changes

- Added filter and sort function for NFTs
- Added filter and sort functions for collections

# Tests

Added tests

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>