feat(frontend): Add data type to ICP wallet worker (#8316)
# Motivation

We are going to onboard more tokens with standard `icp` (as defined in
the the codebase). That means that we need to generalize all the
services and workers for this type of standard, since right now they are
"hard-coded" to be used only by ICP token.

In this PR, we create the type of data to be passed to the worker, that
includes the index canister ID. In a follow-up, we will use such data.

# Changes

- Create type `PostMessageDataRequestIcp` similar to the ones used by
ICRC and DIP20, passing the index canister ID.

# Tests

All tests should work as usual.