fix(frontend): Header z-index with popover backdrops (#8340)
# Motivation

There is an issue where elements of the `StickyHeader` component will
overlap backdrops and menus on the main `Header` component.

This is because the StickyHeaders usually have some popovers that will
render as a BottomSheet (with backdrops) on smaller screens, and these
should overlap everything. Currently the Sticky headers have a `z-index:
3`, as does the main Header.

Since the Header is before the Sticky Headers in the DOM tree, they fall
behind the StickyHeaders backdrops, which is good and only a problem
once we open up Menus/Dropdowns with their own Backdrop that are
positioned within the header, as these fall also behind the
StickyHeaders.

The absolute positioning nature of
Backdrops/Modals/Popovers/StickyHeaders means its impossible with CSS
only to solve this, as some elements are always behind one another,
while the relative positioning of the wrapping elements (in this case
the `Header` and the `StickyHeader` wrapper) create a [stacking
context](https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_positioned_layout/Stacking_context)
that we cannot break out of. (More precisely we can once by using
`position: fixed`, but then we set the priority for that particular
element and always overflow the other one - no solution to this either).

# Changes

To solve the afformentioned issue, we add bindable visible states to all
menus and dropdowns from the main Header and change the headers z-index
depending if one of those is visible/opened.

If none of the menus are open, we keep `z-index: 3` as before, while we
set a `z-index: 4` when a menu has been opened to flow over the sticky
headers.

# Tests

<!-- Please provide any information or screenshots about the tests that
have been done -->

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>