feat(frontend): Post index canister data to ICP wallet worker (#8342)
# Motivation

We are going to onboard more tokens with standard `icp` (as defined in
the the codebase). That means that we need to generalize all the
services and workers for this type of standard, since right now they are
"hard-coded" to be used only by ICP token.

Following up PR https://github.com/dfinity/oisy-wallet/pull/8316, we are
starting to pass the index canister data through postMessage to the ICP
wallet worker.

# Changes

- Pass index canister ID in the message when posting to the ICP wallet
worker.
- Adapt function getTransactions to accept the index canister ID.
- Since for now, we still do not generalize for any icp-standard token,
we hard-code it for ICP token index canister.

# Tests

Current tests are sufficient, once adapted.