feat(frontend): created newSwapForm (#8369)
# Motivation

> **Note:** This component will be renamed and the old `SwapForm` will
be deleted after all follow-up PRs are merged. For now, both need to
coexist to keep the swap logic working.

We need to support multiple networks in Swap.  
This PR introduces a new `NewSwapForm` component that behaves like the
previous `SwapForm`, but all dependencies are passed as **snippets** or
**props**, making it more flexible and easier to extend.

# Changes

- Added new `NewSwapForm` component
- Kept behavior consistent with the old `SwapForm`

# Tests

Covered new Component with tests.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Denys Karmazyn <denys.karmazyn@dfinity.org>