feat(frontend): Guarantee migration of UserTokens to CustomTokens (#8372)
# Motivation

We are migrating from `UserToken` to `CustomToken` for ERC20. We already
set up a few functions and flows for a smooth migration.

However, just to guarantee that no custom token is lost (especially the
ones imported by users long time ago), we create an additional step.

When saving the new ERC20 custom tokens, we will use the existing list
of user tokens as base. And after that we add, uniquely, the modified
ones.