feat(frontend): Use existing metadata for ICP-standard tokens (#8397)
# Motivation

For completeness, we use the existing hard-coded tokens to fetch the
metadata for ICP-standard custom tokens, similar to the other standard.

It is not typically the case that an user adds an ICP-token before we
hard-code it, but just to be sure, we add this feature to avoid fetching
the metadata using a call.