feat(frontend): Markdown utils (#8438)
# Motivation

For the new Terms of Use, License Agreement etc were going to use a
Markdown syntax to store/render the documents.

In order to be able to display a sidebar with links to the main
headings, we need to split it the markdown and mark headings to display
the table of contents.

# Changes

Added utils for getting blocks for a markdown string

# Tests

Added tests