feat(frontend): added new swap modal.svelte component (#8490)
# Motivation

As part of the integration with Velora, a new parent component was added
to encapsulate the logic from the old SwapModal.svelte along with
adjustments for the new flow.

Note: The old SwapModal.svelte component has not been removed yet, to
make rollback easier if something goes wrong.

# Changes
Added NewSwapModal.svelte component