feat(frontend): enable parallel transactions for BTC send (#8455)
# Motivation

This Pull requests enables the parallel execution of BTC transactions
for UTXO's that are currently not locked for the **BTC Send** .

# Changes

- Removed evaluationof the $hasPendingTransactionsStore for disabling
the send button
- Introduced a `debounce` mechanism in `BtcUtxosFee.svelte` to avoid
unnecessary repeated API calls.
- Added a message boxes for all missing `BtcPrepareSendError` and
`BtcSendValidationError` errors and removed the toasts for a better user
experience
- Defined the utxosFee variable as reactive in
`BtcSendTokenWizard.svelte` so it correctly gets evaluated in several
reactive conditions (e.g. disableSend)

# Tests

- Fixed async call issues for ` vi.advanceTimersByTimeAsync()` in
`BtcUtxosFee.spec.ts`
- Aligned unit test `BtcSendReview.spec.ts` with the code changes

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>