feat(backend)!: Add image source flag to `ErcToken` structure (#8336)
# Motivation

We want to allow the user to choose to trust the source of images of a
specific token, being it either for NFT images or icon/logo. So we add
an optional flag to the `ErcToken` structure representing 3 states: not
yet decided (`null`), allowed (`true`), not allowed (`false`).

BREAKING CHANGE: the interface of `ErcToken` is updated.

# Changes

- Add `allow_media_source` field to structure `ErcToken`.

# Tests

Adapted current tests.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>