feat(backend)!: Add agreements data to user profile (#8515)
# Motivation

We will require the user to accept some agreements to use OISY (for
example, License Agreement or Terms of Use).

And we want to store if and when the user accepts (with the timestamp of
each agreement version).

BREAKING CHANGE: The structure of `UserProfile` is being modified, with
a new field for the agreements.

# Changes

- Create new Agreement-related structures in the backend.
- Add a new field to the user profile.

# Tests

Current tests should suffice, for now. We will create some more when we
have the methods to modify the new field.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>