feat(build): Open Content Security Policy for all HTTPS content sources (#8548)
# Motivation

We are going to integrate NFT into OISY, However, most of them have
metadata and images that are URL-sourced. Such URLs are quite vary and
diverse.

Since we are currently providing an hard-coded Content Security Policy
for `connect-src` during build, the solution discussed with ProdSec was
to simply open `connect-src` for now, even with the related risk:

> [...] I would go for option 1 [opening `connect-src`] because the
`connect-src` directive mainly protects against data exfiltration in
case an attacker has achieved RCE, but
> 1. it doesn't change the risk profile of the app (data exfiltration is
already possible via img)
> 2. in a scenario where an attacker has achieved RCE, data exfiltration
is a low priority concern
> 3. solutions that reliably prevent data leakage in case of RCE are
complex to implement and the effort is not worth it IMO because of (2)
> a. maybe it is not even possible because you can also leak data
through the normal interfaces OISY needs access to, like the memo field
of a ICRC transaction for example
>
> The implemented consent modal still makes sense (for reasons we
discussed in the other thread) and I would keep them.


Note that for NFT we will create a consent flow so that the user accepts
the loading of data and images.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>