refactor(frontend): Nft allow media util (#8649)
# Motivation

Currently, the `allowExternalContentSources` only has 2 states for the
UI (true = consent given, false = no consent). But we want to display a
different message when the user never reviewed the consent setting
(undefined = same as false but different message).

Thats also how its stored in the backend.

Currently we dont need to refactor the give consent modal, as for
turning it on or off it will be the same, no matter if the user has done
anything yet.

# Changes

Adjusted return type and value of the getAllowMediaForNft method.

---------

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>