fix(frontend): Accept nullish ETH address in WalletConnect (#8660)
# Motivation

The current logic of the WalletConnect initialization is wrong: we stop
it if either one of between Ethereum address of Solana mainnet address
are nullish.

However, an user could have just one of the two networks enabled.

So the only issue for stopping the initialization if BOTH address are
nullish.