build(frontend): ic-js (v74), signer (v1.0.0) and agent-js (v3.2.4) (#8672)
# Motivation

We have released all our tooling related to the upgrade to AgentJS v3.
We also updated the DID files in ic-js and the signer because of
breaking changes in the ICRC-21 standard (which should not affect OISY,
as we only use the generic consent message, which remains unchanged). We
have also started using semver in the signer. Additionally, a fix
released in AgentJS v3.2.4 was required for the ic-mgmt library, which
this repo uses for local deployment purposes.

# Changes

- Update AgentJS v3 (shipped in
[v4.0.2](https://github.com/dfinity/icp-js-core/releases/tag/v4.0.2)
version of the monorepo)
- Update ic-js [v74](https://github.com/dfinity/ic-js/releases/tag/v74)
and per extension
[v73](https://github.com/dfinity/ic-js/releases/tag/v73) as well.
- Update signer
[v1.0.0](https://github.com/dfinity/oisy-wallet-signer/releases/tag/v1.0.0)
and per extension
[v0.4.0](https://github.com/dfinity/oisy-wallet-signer/releases/tag/v0.4.0)
as well.

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>