feat(frontend): Single IDB cache for custom tokens (#8671)
# Motivation

The IDB cache for custom tokens is divided in networks. However, we have
several kind of standard (or backend variant) that are not being saved
correctly: since we filter the backend custom tokens by standard (or
backend variant), for example, when saving ERC721, we overwrite ERC20,
and viceversa.

It is not scalable to have a single IDB store for each variant or type
of token. And it does not make sense to save all networks custom tokens
in each network IDB cache.

Since we have a single service to load tokens from the backend (and
consequently saving them in the IDB cache) and a single service to load
them for all the networks, we can have a single IDB cache for all of
them.

# Changes

- Create method to set all custom tokens in the cache and use it in
service `loadCustomTokensFromBackend`.
- Create method to get all custom tokens in the IDB cache and use it in
service `loadNetworkCustomTokens` when the use of cache is enabled (note
that we need to filter afterwards for the specific network/variant).
- Create methods to delete/clean the IDB cache and use it in the
appropriate service.
- Remove all deprecated IDB methods for the rest of the custom tokens.

# Tests

Tests adapted.