fix(feature): reduce btc calculation calls   (#8667)
# Motivation

We suddenly start to get timeouts for the following canister calls when
sending multiple time BTC's: after a certain time:
1. `signer.btc_caller_balance`
2. `backend.btc_get_pending_transactions`
3. `bitcoin.btc_caller_send`

Since the `btc_get_pending_transactions` makes a canister call to
bitcoin.bitcoin_get_utxos the most likely cause ist that we are hitting
a firewall threshold on the btc or signer canister.

**This PR is a fix to reduce the number of
`btc_get_pending_transactions` calls**

# Changes

1. Changed filter logic in `btc-wallet.scheduler.ts` to only add btc
transactions when the block height changes. This change required the
storage of the latest block height in the wallet store.

# Tests

- Tested fix on fe2

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>