test(frontend): Create fixtures for Solana integration tests (#8698)
# Motivation

As suggested by prodsec, we should have some integration test on Solana
that checks if our mapping/parsing of the Solana/SPL transactions
matches the balance fetched from the provider.

We will create the tests in PR
https://github.com/dfinity/oisy-wallet/pull/5064. However, the tests
could use the Alchemy provider extensively, and it has rate limit for
requests.

It does not really make sense to continuously fetch real data, if we can
take a snapshot now that is more that satisfactory for the tests.

So, we create fixtures for that specific test. In this PR, we create a
script that saves the fixtures (all public data) and we run it, saving
the results.

# Changes

- Create list of addresses and ATA addresses to use for the fixtures and
the tests.
- Create script to save the fixtures:
  - for addresses:
    - transactions
    - signatures
    - SOL balances
  - for ATA addresses:
    - transactions
    - SPL balances

# Tests

We manually ran the script and saved the results.