test(frontend): compare real Solana transactions with total balance (#5064)
# Motivation

We should really check if our parsing of SOL and SPL tokens matches the
correct amount displayed in the UI too.

To do such a verification, we cross-reference them in a test: we first
calculated all the values of the fetched transactions (and their fees),
and compare it with the balance that we fetch in another service.

# Note

To make this test work, we are using Alchemy API, that, nevertheless,
has a request limit. That means that if we run all the test at the same
time, they will all fail since we touch such limit.

So, we give the choice to use the fixtures created in PR
https://github.com/dfinity/oisy-wallet/pull/8698.

# Changes

- Add Alchemy API key from test environment to the test CI workflow for
the frontend.
- Create all the tests for SOL balances and for SPL balances.
- Load the fixtures and mock the respective functions (it is linked to a
custom variable, so that we can opt-out and use real current data if
necessary).

# Tests

The tests work, all of them.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Denys Karmazyn <denys.karmazyn@dfinity.org>