feat(frontend): service to check if an Index canister is sleepy or awake (#5457)
# Motivation

We had a few cases where the Index canister is in a particular state: it
returns zero balance and no transactions, even if the ledger canister
shows a balance. It may happen that if it is low in cycles, it goes in
"power-saving" mode and in a "sleepy" state, where it does not update
its state.

In these cases, however, it does not raise any error, nor we have any
flag that point it out.

We need a workaround to know if we need to use only the Ledger canister
and ignore the Index canister when the latter is "sleepy". One way of
doing it is to compare the number of blocks: if, after a certain amount
of time (enough to run a syncing), the number of blocks synced by the
Index canister is still discordant with the total number of blocks of
the Ledger canister, then we have this particular state.

# Changes

- Expose method to get the Status of the index canister.
- Create a service `isIndexCanisterAwake` to check if an Index canister
is not updating its state (`sleepy` or `awake`):
- if the number of synced blocks is the same as the total number of
blocks of the Ledger canister, than we can suppose it is `awake` (in any
case it is ok);
- if after 5 seconds (enough to sync at least once), the number of
synced blocks is not different from the previous one, then it is
`sleepy` (because it should have least changed).

# Tests

Included tests.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>