feat(backend)!: experimental features (#8721)
# Motivation

We need to extend user's settings with an additional param -
experimental features. At the moment, there will be just one possible
entry for the AI assistant feature. The idea is to make as flexible as
possible, therefore for the data structure we use the same approach as
it's done for the network settings - each experimental feature can be
prefixed in the future with an additional property (e.g. `FeatureBeta`,
`FeatureAlpha`, and so on) similar to the networks (`BitcoinMainnet`,
`BitcoinTestnet`, etc.). The only option at the moment is gonna be
`AiAssistantBeta`.

BREAKING CHANGE: includes new functionality.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>