feat(frontend): Use global object for initializing WalletConnect (#8849)
# Motivation

During the initialization of the WalletConnect object, there are
sometimes issues with retained values from the requestors. For example,
if we initialize it once to try and reconnect, and then re-initiliaze it
when really connecting, some DEXes will see it fail (for example
https://magiceden.io/).

To avoid it, we create a global object to initialize it.