chore(frontend): Add Sprinkles s1e5 configuration (#8893)
# Motivation

We will start with Sprinkles s1e5

# Changes

- add configuration and initial english texts.

Notes:
- the texts are not yet final. We will likely will get an update on some
texts in a later PR, but we want to have the initial config in, so we
can start reviewing them in the app directly.
- We will update the final start date before the release. For now, we
keep a day in the past, so all deployments to staging from the main
branch keep the episode card active.

# Tests

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>