feat(frontend): update design of ShowContactsToolItem (#8888)
# Motivation

We want to update the ShowContactsToolItem component's design - it now
should behave similar to what we have in the Address book modal. Similar
as with the previous implementation, we do not re-use the existing
component but create a separate one since the usecases (and therefore
params and props) are quite different.

<img width="391" height="821" alt="Screenshot 2025-09-15 at 15 56 43"
src="https://github.com/user-attachments/assets/d2ce59ea-1aa1-446a-b87a-18d60d5d2178"
/>

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>