feat(frontend): `ethers` instead of `alchemy-sdk` when formatting numbers (#8900)
# Motivation

In an effort to deprecate `alchemy-sdk`, we replace the usage of
`Utils.formatUnits` of `alchemy-sdk` (that is inherited from the old
version of `ethers`, alias `@ethersproject`) with `formatUnits` from
`ethers/utils`.

The tests are quite extensive

# Changes

- Replace `Utils.formatUnits` with `formatUnits`.
- The parameter for the units does not accept stringified numbers
anymore, so we need to parse it first.

# Tests

The tests for this util are quite extensive, so they should still work
as intended.