refactor(frontend): Nft Card (#8997)
# Motivation

Adding new type to NftCard and refactoring the way we set a different
type. The default type would be used on transaction details later on.

# Changes

Styling changes
Removed selectable prop and made it a new type

# Tests

Added tests to make sure the relevant styles that have been changed stay
the same for existing UIs

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>