fix(frontend): Center the loading message in the signer flow (#9060)
# Motivation

The loading message in the signer flow was not centered anymore because
we use now `AgreementGuard` to wrap the login: this component has a
`<div> `element for fading that was not inheriting the `flex` class.
However, there is not really need for a fading transition here, so we
can remove it.

### Before

<img width="1158" height="1380" alt="Screenshot 2025-09-22 at 21 31 47"
src="https://github.com/user-attachments/assets/d1c7a06f-f542-4dc1-8030-66988101f064"
/>

### After

<img width="1160" height="1388" alt="Screenshot 2025-09-22 at 21 52 07"
src="https://github.com/user-attachments/assets/82610409-81fe-4d89-8fe0-bf9cc47b744e"
/>

### Entire flow


https://github.com/user-attachments/assets/aa7672d6-b722-40a2-a929-6caeb2e7d1c9

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>