feat(frontend): extend AI contact addresses with acceptedTokenStandards (#9074)
# Motivation

We need to extend AI contact addresses with acceptedTokenStandards. This
field is gonna help LLM to understand whether a token can be sent to an
address. It's especially useful in some particular cases: e.g. `Icrcv2`
address can be a principal or an account id - if it's the former, than
both ICP and ICRC tokens can be sent to it, if the latter - only ICP.

The system prompt will be updated in the follow-up PR.