fix(frontend): `LoaderMultipleEthTransactions` should not stop loading (#9091)
# Motivation

Component `LoaderMultipleEthTransactions` stops loading the
transactions, since we track which tokens are already loaded.

This was useful when we had no control over the loading of the
transactions being called multiple times.

However, now we have the `loading` state that fixes this issue, so we
don't need the list of tokens already loaded anymore.