feat(devops): Ignore test folder when checking unused Svelte components (#9106)
# Motivation

The CI that checks for unused Svelte components, ignores the tests: we
want to see if the components are effectively used.

However, it is giving false positive for Svelte components that are
auxiliary and used ONLY in the tests.

For now, we ignore such component in the analysis, and we will follow-up
with a way to check if they are effectively used, and used ONLY in the
tests.