chore(deps): Remove `jimp` (#9165)
# Motivation

While bumping `zod` to v4 we encountered a library that was sort of
causing dependencies issue: `jimp`.

After a quick analysis, it is used only in the E2E tests, that are
either way being skipped for now and under review.

So, to prioritize the maintenance of the repo, we comment such tests and
uninstall `jimp`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>